Converted fetch tool to use longtail instead of desync
It uses the commandline version of longtail. All the basic steps are operational.
It would be nice to move to the C# library version, but then I think the GCS integration needs to be re-done.
