+++ FetchPrebuiltUE4/FetchPrebuiltUE4.csproj
@@ -68,6 +68,6 @@
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
-    <PostBuildEvent>copy "$(SolutionDir)\Tools\desync.exe" "$(TargetDir)"</PostBuildEvent>
+    <PostBuildEvent>copy "$(SolutionDir)\Tools\longtail.exe" "$(TargetDir)"</PostBuildEvent>
   </PropertyGroup>
 </Project>
